+++ src/Shader/PixelRoutine.cpp
@@ -3584,7 +3584,7 @@ namespace sw
 		int pad = 0;        // Count number of texm3x3pad instructions
 		Vector4i dPairing;   // Destination for first pairing instruction
 
-		for(int i = 0; i < shader->getLength(); i++)
+		for(size_t i = 0; i < shader->getLength(); i++)
 		{
 			const Shader::Instruction *instruction = shader->getInstruction(i);
 			Shader::Opcode opcode = instruction->opcode;
@@ -3786,7 +3786,7 @@ namespace sw
 		bool out[4][4] = {false};
 
 		// Create all call site return blocks up front
-		for(int i = 0; i < shader->getLength(); i++)
+		for(size_t i = 0; i < shader->getLength(); i++)
 		{
 			const Shader::Instruction *instruction = shader->getInstruction(i);
 			Shader::Opcode opcode = instruction->opcode;
@@ -3800,7 +3800,7 @@ namespace sw
 			}
 		}
 		
-		for(int i = 0; i < shader->getLength(); i++)
+		for(size_t i = 0; i < shader->getLength(); i++)
 		{
 			const Shader::Instruction *instruction = shader->getInstruction(i);
 			Shader::Opcode opcode = instruction->opcode;
@@ -5740,7 +5740,7 @@ namespace sw
 
 			if(localShaderConstants)   // Constant may be known at compile time
 			{
-				for(int j = 0; j < shader->getLength(); j++)
+				for(size_t j = 0; j < shader->getLength(); j++)
 				{
 					const Shader::Instruction &instruction = *shader->getInstruction(j);
